Change DocumentConverter[Factory] API.
Use an enum to indicate the file type, and pass the output file name
as a parameter to convert().

It will choose the output format depending on the extension of the
received output file, defaulting to PDF for PPTX and HTML otherwise.

As a side-effect, this reactivates the (very eary stage) Docx->PDF
conversor.