DocxHtml: Improve picture size and alignment.
We translate the width from the Docx into HTML width, and we try to
enhance how it looks with regard to the surrounding text, by setting
alignment and float values.

Align and float values have been heuristically obtained, and it's a
known issue that all the images will float to the left regardless of
how they did in the original.