DocxHtml: Fix paragraph alignment being incorrectly overridden.
Due to a bug in POI, XWPFParagraph.getAlignment() returns LEFT when
the value is not set, overriding the value provided by the style
information.

To address this problem, we interrogate the XML structure directly,
to find out if the parameter is present or not. Only if it's present,
we call getAlignment().